+++ onnx/backend/test/case/node/castlike.py
@@ -184,9 +184,7 @@ class CastLike(Base):
                     "x",
                     to_dtype,
                     input_shape,
-                    vals=np_fp32.astype(from_np_dtype)
-                    .astype(np.float32)
-                    .astype(to_np_dtype),
+                    vals=np_fp32.astype(from_np_dtype).astype(to_np_dtype),
                 )
 
             like = make_tensor("like", to_dtype, (0,), vals=[])
